High-Dynamic-Range-Imaging/code/requirements.txt
matplotlib==3.5.2
numpy==1.23.5
opencv_python==4.5.5.62
Pillow==9.4.0
glob2
argparse
python-math
random2